refactor: Rename errors mod
It seems that `error` is more common (e.g. `nom`)
