+++ src/parse/parser.rs
@@ -14,9 +14,9 @@ use crate::{
     build::{App, Arg, ArgSettings},
     mkeymap::KeyType,
     output::{fmt::Colorizer, Help, HelpWriter, Usage},
-    parse::errors::Error as ClapError,
-    parse::errors::ErrorKind,
-    parse::errors::Result as ClapResult,
+    parse::error::Error as ClapError,
+    parse::error::ErrorKind,
+    parse::error::Result as ClapResult,
     parse::features::suggestions,
     parse::{ArgMatcher, SubCommand},
     parse::{Validator, ValueType},
